Move gui classes to dedicated folder
